Projectile, Collisions and Mouse Position changes
Changed the way the game detect mouse position with a floor that covers the whole map and the character will look correctly without changing orientation depending if the mouse was on a wall or floor. It now detects only the floor.

Projectile was having some issues with collision profiles - Fixed.